% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{accuracy}
\alias{accuracy}
\title{mean accuracy of a classifier}
\usage{
accuracy(real, pred)
}
\arguments{
\item{y}{real value}

\item{yhat}{predicted value}
}
\description{
mean accuracy of a classifier
}
