% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{boosting}
\alias{boosting}
\title{boosting model template/single model}
\usage{
boosting(
  train,
  test,
  formula = y ~ . - id,
  n_trees = 10,
  distribution = c("gaussian", "bernoulli", "multinomial", "poisson", "coxph")
)
}
\arguments{
\item{train}{train dataset}

\item{test}{test dataset}

\item{formula}{gbm::gbm formula}

\item{n_trees}{gbm::gbm n.trees}

\item{distribution}{gbm::gbm distribution}

\item{...}{boosting parameters}
}
\description{
boosting model template/single model
}
