% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosting.R
\name{boosting_maker}
\alias{boosting_maker}
\title{boosting model constructor}
\usage{
boosting_maker(...)
}
\arguments{
\item{...}{boosting parameters}
}
\description{
boosting model constructor
}
\examples{
boostings <- boosting_maker("formula" = "y ~ . - id",
                            "n_trees"  = c(10, seq(from = 25, to = 200, by = 25)),
                            "distribution" = "gaussian")
}
