% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_submission}
\alias{export_submission}
\title{Export kaggle competition submission}
\usage{
export_submission(
  df,
  output_names = c("Id", "SalePrice"),
  input_names = c("id", "y"),
  outdir = "submissions",
  outfile = "submission",
  tg_bot = NULL
)
}
\arguments{
\item{df}{data.frame exported}

\item{output_names}{column names in the export file}

\item{input_names}{column names to consider in the df data.frame}

\item{outfile}{file name}

\item{tg_bot}{telegram::TGBot object already set up with default
chat_id where the file will be posted as well}

\item{directory}{where to export}
}
\description{
Export kaggle competition submission
}
