% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute.R
\name{impute_bag}
\alias{impute_bag}
\title{Impute using bagged trees}
\usage{
impute_bag(df, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with missing variables to be imputed}

\item{ignore}{character vector of ignored variables in imputation}
}
\description{
Impute using bagged trees
}
\examples{
head(airquality)
head(bag_impute(airquality))
}
