% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute.R
\name{impute_knn}
\alias{impute_knn}
\title{imputation using knn (it scales everything)}
\usage{
impute_knn(df, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with missing variables to be imputed}

\item{ignore}{character vector of ignored variables in imputation}
}
\description{
imputation using knn (it scales everything)
}
\examples{
head(airquality)
head(impute_knn(airquality))
}
