% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute.R
\name{impute_median}
\alias{impute_median}
\title{Impute univariately using the median}
\usage{
impute_median(df, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with missing variables to be imputed}

\item{ignore}{character vector of ignored variables in imputation}
}
\description{
Impute univariately using the median
}
\examples{
head(airquality)
head(impute_median(airquality))
}
