% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute.R
\name{impute_mice_long}
\alias{impute_mice_long}
\title{imputation using mice long version (nrow increases)}
\usage{
impute_mice_long(df, m = 5, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with missing variables to be imputed}

\item{m}{the number of imputation to rbind}

\item{ignore}{character vector of ignored variables in imputation}
}
\description{
imputation using mice long version (nrow increases)
}
\examples{
dim(airquality)
imp <- impute_mice_long(airquality)
dim(imp)
}
