% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_impute.R
\name{impute_mice_single}
\alias{impute_mice_single}
\title{imputation using mice (one imputation)}
\usage{
impute_mice_single(df, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with missing variables to be imputed}

\item{ignore}{character vector of ignored variables in imputation}
}
\description{
imputation using mice (one imputation)
}
\examples{
head(airquality)
head(impute_mice_single(airquality))
}
