% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbml-package.R
\docType{package}
\name{lbml-package}
\alias{lbml}
\alias{lbml-package}
\title{lbml: ML wrappers and boilerplate code}
\description{
Miscellaneous machine learning wrappers and boilerplate code.
}
\author{
\strong{Maintainer}: Luca Braglia \email{lbraglia@gmail.com}

}
\keyword{internal}
