% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{lm_maker}
\alias{lm_maker}
\title{construct several predictor given formula}
\usage{
lm_maker(formulas_list)
}
\arguments{
\item{formulas_list}{list of formulas}
}
\description{
construct several predictor given formula
}
\examples{
formulas <- list("all variables" = y ~ . - id)
lms <- lm_maker(formulas)
}
