% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{logistic}
\alias{logistic}
\title{random forest template predictors (quntitative)}
\usage{
logistic(train, test, formula = y ~ . - id)
}
\arguments{
\item{train}{training dataset}

\item{test}{test dataset}

\item{formula}{formula used in the randomForest}
}
\description{
random forest template predictors (quntitative)
}
\examples{
logistic(house_train, house_test, y ~ Pclass + Age)
}
