% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{logistic_maker}
\alias{logistic_maker}
\title{create logistics models}
\usage{
logistic_maker(formula_list)
}
\arguments{
\item{formula_list}{a list of formulas to use in rf}
}
\description{
create logistics models
}
\examples{
formulas <- list("mod1" = y ~ Pclass + Age,
                "mod2" = y ~ Pclass + Age + SibSp + Parch + Embarked)
logistics <- logistic_maker(formulas)
logistics[[1]](titanic_train, titanic_test)
logistics[[2]](titanic_train, titanic_test)
}
