% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{naive_bayes_g}
\alias{naive_bayes_g}
\title{Naive bayes (gaussian) predictor}
\usage{
naive_bayes_g(train, test)
}
\description{
Naive bayes (gaussian) predictor
}
\examples{
## naive_bayes_g(algo_train, algo_test)# it's a NA problem
## naive_bayes_g(median_impute(algo_train), median_impute(algo_test))
}
