% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{naive_bayes_k}
\alias{naive_bayes_k}
\title{Naive bayes (kernel) predictor}
\usage{
naive_bayes_k(train, test)
}
\description{
Naive bayes (kernel) predictor
}
\examples{
# naive_bayes_k(median_impute(algo_train), median_impute(algo_test))
}
