% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_cap}
\alias{outliers_cap}
\title{cap the outliers}
\usage{
outliers_cap(df, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with possible outliers to be capped}

\item{ignore}{character vector of ignored variable in imputation}
}
\description{
cap the outliers
}
\examples{
x <- airquality
x[1,1] <- 1e06
head(outliers_cap(x))
}
