% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_identifier_iqr}
\alias{outliers_identifier_iqr}
\title{identify outliers based on iqr}
\usage{
outliers_identifier_iqr(x)
}
\arguments{
\item{x}{a numeric variable}
}
\description{
identify outliers based on iqr
}
