% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_identifier_z}
\alias{outliers_identifier_z}
\title{identify outliers based on z scores}
\usage{
outliers_identifier_z(x)
}
\arguments{
\item{x}{a numeric variable}
}
\description{
identify outliers based on z scores
}
