% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_none}
\alias{outliers_none}
\title{Ignore the outliers}
\usage{
outliers_none(df, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with possible outliers}

\item{ignore}{character vector of ignored variable in imputation}
}
\description{
it returns the dataframe unchanged
}
