% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers_trim_iqr}
\alias{outliers_trim_iqr}
\title{trim observations with outliers based on IQR}
\usage{
outliers_trim_iqr(df, ignore = c("y", "id"))
}
\arguments{
\item{df}{a data.frame with possible outliers to be capped}

\item{ignore}{character vector of ignored variable in imputation}
}
\description{
trim observations with outliers based on IQR
}
\examples{
x <- airquality
x[1,1] <- 1e06
head(outliers_trim_iqr(x))
}
