% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest}
\alias{random_forest}
\title{random forest template predictors (quntitative)}
\usage{
random_forest(train, test, formula = y ~ . - id)
}
\arguments{
\item{train}{training dataset}

\item{test}{test dataset}

\item{formula}{formula used in the randomForest}
}
\description{
random forest template predictors (quntitative)
}
\examples{
random_forest(house_train, house_test)
}
