% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{random_forest_maker}
\alias{random_forest_maker}
\title{create random forest}
\usage{
random_forest_maker(formula_list)
}
\arguments{
\item{formula_list}{a list of formulas to use in rf}
}
\description{
create random forest
}
\examples{
formulas <- list("mod1" = y ~ .)
rfs <- random_forest_maker(formulas)
rfs[[1]](house_train, house_test)
}
