% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{standardize_df}
\alias{standardize_df}
\title{standardize data frames}
\usage{
standardize_df(
  train,
  test,
  id_var = "Id",
  outcome_var = "SalePrice",
  return_splitted = TRUE
)
}
\arguments{
\item{train}{train dataset}

\item{id_var}{id column in both the dataset}

\item{outcome_var}{outcome column in both the dataset}
}
\description{
standardize data frames
}
