% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{svm}
\alias{svm}
\title{Support vector machine predictor}
\usage{
svm(train, test, formula = y ~ . - id, kernel = "radial", cost = 1)
}
\description{
Support vector machine predictor
}
\examples{
## svm(algo_train, algo_test)
}
