% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{svm_maker}
\alias{svm_maker}
\title{Support vector machines maker}
\usage{
svm_maker(...)
}
\description{
Support vector machines maker
}
\examples{
svms <- svm_maker("formula" = "y ~ . - id",
                  "kernel"  = c("linear", "radial", "polynomial"),
                  "cost" = 10^{-2:2})
## svms[[1]](algo_train, algo_test)
## svms[[9]](algo_train, algo_test)
}
